More Character.js refactoring
