+++ module/rolls/rollStat.ts
@@ -1,4 +1,4 @@
-import { Ability, TracksTests, BWCharacter } from "../bwactor.js";
+import { Ability, TracksTests, BWCharacter, BWActor } from "../bwactor.js";
 import { BWActorSheet } from "../bwactor-sheet.js";
 import * as helpers from "../helpers.js";
 import {
@@ -16,7 +16,7 @@ import {
 
 export async function handleStatRoll({ target, sheet }: RollOptions): Promise<unknown> {
     const stat = getProperty(sheet.actor.data, target.dataset.accessor || "") as Ability;
-    const actor = sheet.actor as BWCharacter;
+    const actor = sheet.actor as BWActor & BWCharacter;
     const statName = target.dataset.rollableName || "Unknown Stat";
     const rollModifiers = sheet.actor.getRollModifiers(statName);
     let tax = 0;
@@ -95,7 +95,7 @@ async function statRollCallback(
         callons
     };
     if (sheet.actor.data.type === "character") {
-        (sheet.actor as BWCharacter).addStatTest(stat, name, accessor, dg, isSuccessful);
+        (sheet.actor as BWActor & BWCharacter).addStatTest(stat, name, accessor, dg, isSuccessful);
     }
 
     const messageHtml = await renderTemplate(templates.skillMessage, data);
